requirements.txt
requests
openai==0.28
python-dotenv